requirements.txt
asgiref==3.11.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.4
dj-database-url==3.1.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==24.1.1
h11==0.16.0
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.10.1
python-dotenv==1.2.1
python-http-client==3.3.7
sendgrid==6.12.5
sqlparse==0.5.5
uvicorn==0.40.0
Werkzeug==3.1.5
whitenoise==6.11.0
